Refer RB for problem description and logs
Fix is to not take lock while deleting the hostif. Though this will avoid deadlock,
in rare situation we might crash(While we delete the hostif, the packet processing
function, process_packet_for_fdb_event which is in another thread might still be
usingthat hostif)
But all this has been solved by refactoring the code in github master and 202006 branch.
But right now, its too complicated and takes significant effort to bring those
changes in 201911. Also keeping in mind that this happens only in VS environment,
we are moving ahead with this incomplete solution for now.

RB=2220296
G=lnos-reviewers
R=pchaudhary,pmao,samaity,zxu
A=zxu